+++ fields/types/list/ListField.js
@@ -5,7 +5,7 @@ import { css, StyleSheet } from 'aphrodite/no-important';
 import React from 'react';
 import Field from '../Field';
 import Domify from 'react-domify';
-import _map from 'lodash/map';
+import _cloneDeep from 'lodash/cloneDeep';
 
 import { Fields } from 'FieldTypes';
 import { Button, GlyphButton } from '../../../admin/client/App/elemental';
@@ -16,7 +16,8 @@ function generateId () {
 	return i++;
 };
 
-const ItemDom = ({ name, id, onRemove, children, t }) => (
+
+const ItemDom = ({  name, id, onRemove, children, t }) => (
 	<div style={{
 		borderTop: '2px solid #eee',
 		paddingTop: 15,
@@ -27,6 +28,7 @@ const ItemDom = ({ name, id, onRemove, children, t }) => (
 			return React.cloneElement(child, {
 				name,
 				id,
+				onRemove,
 				t,
 			});
 		})}
@@ -86,13 +88,40 @@ module.exports = Field.create({
 			if (typeof Fields[field.type] !== 'function') {
 				return React.createElement(InvalidFieldType, { type: field.type, path: field.path, key: field.path });
 			}
-			const props = assign({}, field);
+
+			/* 
+			** getRelatedFilter is support single not nested field
+			** so that, clone the list and set subField data to one-level data
+			** Fung Lee
+			** 13/06/2019
+			*/
+			const list = _cloneDeep(this.props.list);
+			list.fields = this.props.list.fields[this.props.path].fields;
+
+			// cant get child state here, so imitate state from parent props data
+			const state = {
+				values: this.props.values[this.props.path][index],
+			};
+
+			let props = assign({}, field);
+			props.list = list;
+			props.path = field.path;
 			props.value = value[field.path];
 			props.values = value;
 			props.onChange = this.handleFieldChange.bind(this, index);
 			props.mode = 'edit';
 			props.inputNamePrefix = `${this.props.path}[${index}]`;
 			props.key = field.path;
+
+			const filters = list.getRelatedFilter(field, props.filters, props, state);
+
+			if (filters) {
+				props = {
+					...props,
+					filters,
+				};
+			}
+
 			// TODO ?
 			// if (props.dependsOn) {
 			// 	props.currentDependencies = {};
@@ -119,6 +148,7 @@ module.exports = Field.create({
 						</ItemDom>
 					);
 				})}
+
 				<GlyphButton color="success" glyph="plus" position="left" onClick={onAdd}>
 					{t('add')}
 				</GlyphButton>
@@ -145,7 +175,8 @@ module.exports = Field.create({
 const classes = StyleSheet.create({
 	container: {
 		marginTop: '2em',
-		paddingLeft: '2em',
-		boxShadow: '-3px 0 0 rgba(0, 0, 0, 0.1)',
+		marginLeft: '1em',
+		paddingLeft: '1em',
+		boxShadow: '-2px 0 0 rgba(0, 0, 0, 0.1)',
 	},
 });